+++ vivarium/experiments/large_database_experiment.py
@@ -3,8 +3,9 @@ Experiment to test maximum BSON document size with MongoDB emitter
 """
 
 import random
-from vivarium.core.experiment import Experiment
-from vivarium.core.process import Process, Composer
+from vivarium.core.engine import Engine
+from vivarium.core.process import Process
+from vivarium.core.composer import Composer
 from vivarium.core.emitter import (
     get_experiment_database,
     data_from_database,
@@ -68,7 +69,7 @@ def run_large_initial_emit():
         'emitter': 'database'
     }
 
-    experiment = Experiment({
+    experiment = Engine({
         'processes': composite['processes'],
         'topology': composite['topology'],
         **settings})