+++ vivarium/processes/remove.py
@@ -7,11 +7,11 @@ Remove Process
 import os
 from typing import Any, Dict
 
-from vivarium.core.experiment import pp
+from vivarium.core.engine import pp
 from vivarium.core.process import (
     Deriver,
-    Composer,
 )
+from vivarium.core.composer import Composer
 from vivarium.core.composition import (
     simulate_composer,
     PROCESS_OUT_DIR,