+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+
+## v0.3.0
+
+* Store API supports simplified scripting of a bigraph
+* Breaking import changes: `Composite` and `Composer` are now under `vivarium.core.composer`, and `Experiment` has been renamed `Engine` under `vivarium.core.engine`.
+
+
 ## v0.2.16
 
 * add more flexible kwarg passing in `Control`